+++ mrp_bom_variable/models/__init__.py
@@ -1,1 +1,2 @@
 from . import mrp_bom
+from . import mrp_bom_line
